+++ packages/engine/src/index.ts
@@ -1,4 +1,4 @@
-export * from './HedronEngine'
+export * from './HedronEngine/HedronEngine'
 export type * from '@store/types'
 export * from '@store/types'
 import './globalVars'
